Use unique names for temporary zip files
When a zip archive was created to fill a download request, it was
created at a hardcoded path ('/tmp/download.zip') which caused
a bunch of problems. Now the filename is a SHA-1 hash of the list
of requested song IDs and the file is removed after the request
is fulfilled.