Implements cleaner, more reliable shutdown
Added a few things to make sure that shutting down goes more
smoothly and we don't end up with lingering processes or threads.
Bonus side effect is now we can cancel a library update. Bonus
regression is that non-ASCII characters in filenames won't throw
exceptions when attempting to transcode them, and the process
just hangs.